pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 64"]

[project]
name = "data_explorer_agent"
version = "0.1.0"
dependencies = [
  "nvidia-nat[langchain]~=1.3",
]
requires-python = ">=3.11,<3.14"
description = "Custom NeMo Agent Toolkit Workflow"
classifiers = ["Programming Language :: Python"]



[project.entry-points.'nat.components']
data_explorer_agent = "data_explorer_agent.register"